add request response log operation controller & view

+++ FamilyTree_Backend/InternalServices/FamilyTreeService.cs
@@ -454,7 +454,7 @@ namespace FamilyTreeBackend.Infrastructure.Service.InternalServices
                 .Where(tr => tr.DateCreated == null || tr.DateCreated.Value.CompareTo(model.CreatedBefore) <= 0)
                 .Where(tr => tr.PublicMode == true);
 
-            var totalPage = (ulong)MathF.Ceiling((ulong)trees.Count() / model.ItemsPerPage);
+            var totalPage = (ulong)MathF.Ceiling(trees.Count() / (float)model.ItemsPerPage);
             totalPage = totalPage <= 0 ? 1 : totalPage;
 
             if (model.Page > totalPage)
@@ -493,7 +493,7 @@ namespace FamilyTreeBackend.Infrastructure.Service.InternalServices
                 .Where(tr => tr.DateCreated == null || tr.DateCreated.Value.CompareTo(model.CreatedBefore) <= 0)
                 .Where(tr => tr.OwnerId.Equals(applicationUser.Id) || tr.Editors.Any(e => e.Id.Equals(applicationUser.Id)));
 
-            var totalPage = (ulong)MathF.Ceiling((ulong)trees.Count() / model.ItemsPerPage);
+            var totalPage = (ulong)MathF.Ceiling((ulong)trees.Count() / (float)model.ItemsPerPage);
             totalPage = totalPage <= 0 ? 1 : totalPage;
 
             if (model.Page > totalPage)
@@ -536,7 +536,7 @@ namespace FamilyTreeBackend.Infrastructure.Service.InternalServices
             query = query.Where(tr => tr.DateCreated == null || tr.DateCreated.Value.CompareTo(model.CreatedBefore) <= 0)
                 .Where(tr => tr.PublicMode == true);
 
-            var totalPage = (ulong)MathF.Ceiling((ulong)query.Count() / model.ItemsPerPage);
+            var totalPage = (ulong)MathF.Ceiling((ulong)query.Count() / (float)model.ItemsPerPage);
             totalPage = totalPage <= 0 ? 1 : totalPage;
 
             if (model.Page > totalPage)
@@ -588,7 +588,7 @@ namespace FamilyTreeBackend.Infrastructure.Service.InternalServices
             query = query.Where(tr => tr.DateCreated == null || tr.DateCreated.Value.CompareTo(model.CreatedBefore) <= 0)
                 .Where(tr => tr.PublicMode == true || tr.OwnerId.Equals(applicationUser.Id) || tr.Editors.Any(e => e.Id.Equals(applicationUser.Id)));
 
-            var totalPage = (ulong)MathF.Ceiling((ulong)query.Count() / model.ItemsPerPage);
+            var totalPage = (ulong)MathF.Ceiling((ulong)query.Count() / (float)model.ItemsPerPage);
             totalPage = totalPage <= 0 ? 1 : totalPage;
 
             if (model.Page > totalPage)
